fix entity relations between characterInventory and equipItem / change id data type from Biginteger to Long

+++ src/main/java/com/jin/game/repository/ItemRepository.java
@@ -3,7 +3,6 @@ package com.jin.game.repository;
 import com.jin.game.model.items.GameItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 
-public interface ItemRepository extends JpaRepository<GameItem, BigInteger> {
+public interface ItemRepository extends JpaRepository<GameItem, Long> {
 }
